// src/app/dashboard/models/unbilled-revenue-filter.model.ts
/*
 * Created on Wed Aug 28 2019
 * Created by - 1157: Mayur Ranjan
 *
 * Copyright (c) 2019 Fortigo Network Logistics Pvt Ltd (4TiGO)
 */

export class UnbilledRevenueFilterModel {
    fromDate: string;
    toDate: string;
    dateFlag: string;
    rmIds: Array<string>;
    sourcingMgrIds: Array<string>;
    companyIds: Array<string>;
    collectionFfeIds: Array<string>;
}